fix(chat): refactor error boundary wrapper
